requirements.txt
# Combined requirements for Lambda functions
# Includes development dependencies (ruff, pytest)
boto3==1.37.38
psycopg==3.2.4
psycopg-binary==3.2.4
pytest==7.4.3
pytest-watch==4.2.0
python-dotenv==1.0.1
requests==2.32.3
ruff==0.3.4
twilio==9.4.1
